+++ libtelemetry/src/main/java/com/mapbox/android/telemetry/TelemetryClientSettings.java
@@ -1,6 +1,8 @@
 package com.mapbox.android.telemetry;
 
 
+import android.support.annotation.Nullable;
+
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
@@ -11,6 +13,7 @@ import javax.net.ssl.X509TrustManager;
 
 import okhttp3.ConnectionSpec;
 import okhttp3.HttpUrl;
+import okhttp3.Interceptor;
 import okhttp3.OkHttpClient;
 
 class TelemetryClientSettings {
@@ -47,8 +50,12 @@ class TelemetryClientSettings {
     return environment;
   }
 
-  OkHttpClient getClient() {
-    return configureHttpClient();
+  OkHttpClient getClient(CertificateBlacklist certificateBlacklist) {
+    return configureHttpClient(certificateBlacklist, new GzipRequestInterceptor());
+  }
+
+  OkHttpClient getAttachmentClient(CertificateBlacklist certificateBlacklist) {
+    return configureHttpClient(certificateBlacklist, null);
   }
 
   HttpUrl getBaseUrl() {
@@ -136,13 +143,18 @@ class TelemetryClientSettings {
     }
   }
 
-  private OkHttpClient configureHttpClient() {
+  private OkHttpClient configureHttpClient(CertificateBlacklist certificateBlacklist,
+                                           @Nullable Interceptor interceptor) {
     CertificatePinnerFactory factory = new CertificatePinnerFactory();
     OkHttpClient.Builder builder = client.newBuilder()
-      .addInterceptor(new GzipRequestInterceptor())
       .retryOnConnectionFailure(true)
-      .certificatePinner(factory.provideCertificatePinnerFor(environment))
+      .certificatePinner(factory.provideCertificatePinnerFor(environment, certificateBlacklist))
       .connectionSpecs(Arrays.asList(ConnectionSpec.MODERN_TLS, ConnectionSpec.COMPATIBLE_TLS));
+
+    if (interceptor != null) {
+      builder.addInterceptor(interceptor);
+    }
+
     if (isSocketFactoryUnset(sslSocketFactory, x509TrustManager)) {
       builder.sslSocketFactory(sslSocketFactory, x509TrustManager);
       builder.hostnameVerifier(hostnameVerifier);