+++ CHANGELOG.md
@@ -4,6 +4,29 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## Mapbox Android Telemetry
 
+### v3.2.0
+- Change fieldIds to fileIds - [#212](https://github.com/mapbox/mapbox-events-android/pull/212)
+- Attachment Event - [#205](https://github.com/mapbox/mapbox-events-android/pull/205)
+- Null Checks for obtainSharedPreferences - [#202](https://github.com/mapbox/mapbox-events-android/pull/202)
+- Create General VisionEvent - [#197](https://github.com/mapbox/mapbox-events-android/pull/197)
+- Add Maps UserAgent - [#179](https://github.com/mapbox/mapbox-events-android/pull/179)
+- New Hashes and Blacklist - [#151](https://github.com/mapbox/mapbox-events-android/pull/151)
+
+### v3.1.5
+- Check Application Context when stopping service - [#190](https://github.com/mapbox/mapbox-events-android/pull/190)
+- Crash Fix When Disabling Location Permissions Manually - [#184](https://github.com/mapbox/mapbox-events-android/pull/184)
+- Downgrade ArchLifecycleVersion - [#183](https://github.com/mapbox/mapbox-events-android/pull/183)
+- Reduce API level for Current State check - [#180](https://github.com/mapbox/mapbox-events-android/pull/180)
+
+### v3.1.4
+- Adjust Wakeup - [#173](https://github.com/mapbox/mapbox-events-android/pull/173)
+- Remove MAS Dependency - [#164](https://github.com/mapbox/mapbox-events-android/pull/164)
+
+### v3.1.3
+- Oreo Background Service Crash Fix - [#157](https://github.com/mapbox/mapbox-events-android/pull/157)
+- TelemetryLocationEnabler Null Bug Fix - [#156](https://github.com/mapbox/mapbox-events-android/pull/156)
+- Make LocationEvent constructor Package-Private - [#155](https://github.com/mapbox/mapbox-events-android/pull/155)
+
 ### v3.1.2
 - Service Stability Fix - [#152](https://github.com/mapbox/mapbox-events-android/pull/152)
 
@@ -77,6 +100,10 @@ Mapbox welcomes participation and contributions from everyone.
 - Port Location Code from MAS - [#6](https://github.com/mapbox/mapbox-events-android/pull/6)
 
 ## Mapbox Android Core
+
+### v0.2.1
+- Update Proguard Rules for GMS - [#189](https://github.com/mapbox/mapbox-events-android/pull/189)
+- GoogleLocationEngine lack of priority crash Fix - [#178](https://github.com/mapbox/mapbox-events-android/pull/178)
 
 ### v0.2.0
 - FastestInterval Documentation Update - [#111](https://github.com/mapbox/mapbox-events-android/pull/111)